# Cleaned emitter to not emit preamble if unneeded