Did the following experiment to check if there was any performance impact on rocket.fir:

New emitter
2137.1 ms
2133.6 ms
2051.0 ms
2166.3 ms

Old emitter
2009.8 ms
1556.9 ms
2030.8 ms
1596.6 ms

Looks like they are about the same, and small, so I think its good.