@azidar 2017-08-24
Addressed comments, thanks!