@azidar 2017-08-25
Redid to emit preamble before initializations, not at top of file. Still need to check whether it works with rocketchip