@azidar 2017-09-06
Works with rocketchip, can merge.